+++ plone/recipe/codeanalysis/tests/test_docs.py
@@ -5,7 +5,6 @@ import doctest
 import os
 import unittest
 import zc.buildout.testing
-import zc.buildout.tests
 
 
 optionflags = (